fix the usage of Authorizer in IoTDB

+++ src/main/java/cn/edu/tsinghua/iotdb/qp/executor/OverflowQPExecutor.java
@@ -3,6 +3,7 @@ package cn.edu.tsinghua.iotdb.qp.executor;
 import cn.edu.tsinghua.iotdb.auth.AuthException;
 import cn.edu.tsinghua.iotdb.auth.authorizer.IAuthorizer;
 import cn.edu.tsinghua.iotdb.auth.authorizer.LocalFileAuthorizer;
+import cn.edu.tsinghua.iotdb.auth.entity.PrivilegeType;
 import cn.edu.tsinghua.iotdb.engine.filenode.FileNodeManager;
 import cn.edu.tsinghua.iotdb.exception.ArgsErrorException;
 import cn.edu.tsinghua.iotdb.exception.FileNodeManagerException;
@@ -325,6 +326,7 @@ public class OverflowQPExecutor extends QueryProcessExecutor {
 		String roleName = author.getRoleName();
 		String password = author.getPassword();
 		String newPassword = author.getNewPassword();
+		String proposer = author.getProposer();
 		Set<Integer> permissions = author.getPermissions();
 		Path nodeName = author.getNodeName();
 		IAuthorizer authorizer = LocalFileAuthorizer.getInstance();
@@ -332,51 +334,72 @@ public class OverflowQPExecutor extends QueryProcessExecutor {
 			boolean flag = true;
 			switch (authorType) {
 			case UPDATE_USER:
-				authorizer.updateUserPassword(userName, newPassword);
+				if(!proposer.equals(userName))
+					if(!authorizer.checkUserPrivileges(userName, null, PrivilegeType.MODIFY_PASSWORD.ordinal()))
+						throw new ProcessorException("No privilege to " + authorType);
+				return authorizer.updateUserPassword(userName, newPassword);
 			case CREATE_USER:
-				authorizer.createUser(userName, password);
+				if(!authorizer.checkUserPrivileges(userName, null, PrivilegeType.CREATE_USER.ordinal()))
+					throw new ProcessorException("No privilege to " + authorType);
+				return authorizer.createUser(userName, password);
 			case CREATE_ROLE:
-				authorizer.createRole(roleName);
+				if(!authorizer.checkUserPrivileges(userName, null, PrivilegeType.CREATE_ROLE.ordinal()))
+					throw new ProcessorException("No privilege to " + authorType);
+				return authorizer.createRole(roleName);
 			case DROP_USER:
+				if(!authorizer.checkUserPrivileges(userName, null, PrivilegeType.DELETE_USER.ordinal()))
+					throw new ProcessorException("No privilege to " + authorType);
 				return authorizer.deleteUser(userName);
 			case DROP_ROLE:
+				if(!authorizer.checkUserPrivileges(userName, null, PrivilegeType.CREATE_ROLE.ordinal()))
+					throw new ProcessorException("No privilege to " + authorType);
 				return authorizer.deleteRole(roleName);
 			case GRANT_ROLE:
+				if(!authorizer.checkUserPrivileges(userName, null, PrivilegeType.GRANT_ROLE_PRIVILEGE.ordinal()))
+					throw new ProcessorException("No privilege to " + authorType);
 				for (int i : permissions) {
 					if (!authorizer.grantPrivilegeToRole(roleName, nodeName.getFullPath(), i))
 						flag = false;
 				}
 				return flag;
 			case GRANT_USER:
+				if(!authorizer.checkUserPrivileges(userName, null, PrivilegeType.GRANT_USER_PRIVILEGE.ordinal()))
+					throw new ProcessorException("No privilege to " + authorType);
 				for (int i : permissions) {
 					if (!authorizer.grantPrivilegeToUser(userName, nodeName.getFullPath(), i))
 						flag = false;
 				}
 				return flag;
 			case GRANT_ROLE_TO_USER:
+				if(!authorizer.checkUserPrivileges(userName, null, PrivilegeType.GRANT_USER_ROLE.ordinal()))
+					throw new ProcessorException("No privilege to " + authorType);
 				return authorizer.grantRoleToUser(roleName, userName);
 			case REVOKE_USER:
+				if(!authorizer.checkUserPrivileges(userName, null, PrivilegeType.REVOKE_USER_PRIVILEGE.ordinal()))
+					throw new ProcessorException("No privilege to " + authorType);
 				for (int i : permissions) {
 					if (!authorizer.revokePrivilegeFromUser(userName, nodeName.getFullPath(), i))
 						flag = false;
 				}
 				return flag;
 			case REVOKE_ROLE:
+				if(!authorizer.checkUserPrivileges(userName, null, PrivilegeType.REVOKE_ROLE_PRIVILEGE.ordinal()))
+					throw new ProcessorException("No privilege to " + authorType);
 				for (int i : permissions) {
 					if (!authorizer.revokePrivilegeFromRole(roleName, nodeName.getFullPath(), i))
 						flag = false;
 				}
 				return flag;
 			case REVOKE_ROLE_FROM_USER:
+				if(!authorizer.checkUserPrivileges(userName, null, PrivilegeType.REVOKE_USER_ROLE.ordinal()))
+					throw new ProcessorException("No privilege to " + authorType);
 				return authorizer.revokeRoleFromUser(roleName, userName);
 			default:
-				break;
-
+				throw new ProcessorException("Unrecognized operator " + authorType);
 			}
 		} catch (AuthException e) {
 			throw new ProcessorException(e.getMessage());
 		}
-		return false;
 	}
 
 	private boolean operateMetadata(MetadataPlan metadataPlan) throws ProcessorException {
